valence_text
