+++ crates/valence_build_utils/src/lib.rs
@@ -38,6 +38,8 @@ pub fn write_generated_file(content: TokenStream, out_file: &str) -> anyhow::Res
     Ok(())
 }
 
+/// Parses a [`proc_macro2::Ident`] from a `str`. Rust keywords are prepended
+/// with underscores to make them valid identifiers.
 pub fn ident(s: impl AsRef<str>) -> Ident {
     let s = s.as_ref().trim();
