+++ crates/valence_text/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "valence_text"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+documentation.workspace = true
+license.workspace = true
+
+[features]
+translate = []
+
+[dependencies]
+anyhow.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+thiserror.workspace = true
+uuid = { workspace = true, features = ["serde"] }
+valence_ident.workspace = true
+valence_nbt.workspace = true
+
+[build-dependencies]
+anyhow.workspace = true
+heck.workspace = true
+proc-macro2.workspace = true
+quote.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+valence_build_utils.workspace = true
