+++ crates/valence_text/build.rs
@@ -1,0 +1,49 @@
+pub fn main() {
+    #[cfg(feature = "translate")]
+    valence_build_utils::write_generated_file(translate::build().unwrap(), "translate.rs").unwrap();
+}
+
+#[cfg(feature = "translate")]
+mod translate {
+    use heck::ToShoutySnakeCase;
+    use proc_macro2::TokenStream;
+    use quote::quote;
+    use serde::Deserialize;
+    use valence_build_utils::ident;
+
+    pub fn build() -> anyhow::Result<TokenStream> {
+        let translations = serde_json::from_str::<Vec<Translation>>(include_str!(
+            "../../extracted/translation_keys.json"
+        ))?;
+
+        let translation_key_consts = translations
+            .iter()
+            .map(|translation| {
+                let const_id = ident(translation.key.replace('.', "_").to_shouty_snake_case());
+                let key = &translation.key;
+                let english_translation = &translation.english_translation;
+                let doc = format!("\"{}\"", escape(english_translation));
+
+                quote! {
+                    #[doc = #doc]
+                    pub const #const_id: &str = #key;
+                }
+            })
+            .collect::<Vec<TokenStream>>();
+
+        Ok(quote! {
+            #(#translation_key_consts)*
+        })
+    }
+
+    #[derive(Deserialize, Clone, Debug)]
+    struct Translation {
+        key: String,
+        english_translation: String,
+    }
+
+    /// Escapes characters that have special meaning inside docs.
+    fn escape(text: &str) -> String {
+        text.replace('[', "\\[").replace(']', "\\]")
+    }
+}
