+++ crates/valence_text/src/translate.rs
@@ -1,0 +1,3 @@
+//! Contains all of Minecraft's vanilla translation keys. 
+
+include!(concat!(env!("OUT_DIR"), "/translate.rs"));
